fix(warnings): resolve shadow/sign-compare/uninit/pedantic C warnings (Phase 1)
- decoration_provider.c: rename `display_tick` parameter to `dtick` in
  nvim_decor_cb_buf/end/start to eliminate -Wshadow (the global
  `display_tick` at globals.h:781 was being shadowed).
- winfloat/src/lib.rs + api/win_config.c: change rs_border_char_count()
  return type from c_int to usize (Rust) / size_t (C extern decl) so
  the comparisons against size_t loop variables are sign-clean (-Wsign-compare).
- api/win_config.c:709: initialize `char *pos` to "left" (the kAlignLeft
  value) so GCC doesn't flag the switch as possibly-uninitialized; all
  three enum arms still overwrite it, no behavior change (-Wmaybe-uninitialized).
- macros_defs.h: add PRAGMA_DIAG_PUSH_IGNORE_PEDANTIC macro alongside the
  existing PUSH_IGNORE_MISSING_PROTOTYPES/IMPLICIT_FALLTHROUGH family, with
  GCC, Clang, and empty-fallback variants.
- eval_shim.c: wrap the irreducible function-pointer→void* cast in
  nvim_chan_proc_set_exit_cb with PRAGMA_DIAG_PUSH_IGNORE_PEDANTIC /
  PRAGMA_DIAG_POP (-Wpedantic). Documented as an FFI boundary constraint.

+++ src/nvim-rs/winfloat/src/lib.rs
@@ -431,8 +431,8 @@ pub const BORDER_CHAR_COUNT: usize = 8;
 
 /// Get number of border characters
 #[no_mangle]
-pub extern "C" fn rs_border_char_count() -> c_int {
-    BORDER_CHAR_COUNT as c_int
+pub extern "C" fn rs_border_char_count() -> usize {
+    BORDER_CHAR_COUNT
 }
 
 // =============================================================================
